D3mPipelineSearch/requirements.txt
click==6.7
enum34==1.1.6
futures==3.1.1
grpcio==1.10.1
grpcio-tools==1.10.1
MarkupSafe==1.0
pathlib==1.0.1
protobuf==3.5.2.post1
six==1.11.0
Werkzeug==0.14.1
